feat(entities): denormalize entity subtypes
Entity subtypes have been denormalized. Subtype IDs are no longer
maintained.
Subtypes are stored as raw arbitrary strings.

BREAKING CHANGE
entity_subtypes database table has been removed.
Entity subtypes no longer have IDs, subsequently all API around
entity subtypes has been removed, including add_subtype(),
update_subtype(), remove_subtype(), get_subtype_from_id(),
get_subtype_id(), get_subtype_class(), get_subtype_class_from_id().
Plugins can define custom entity classes using elgg_set_entity_class()
at runtime.

+++ engine/tests/phpunit/integration/Elgg/Integration/ElggCoreAccessSQLTest.php
@@ -22,7 +22,9 @@ class ElggCoreAccessSQLTest extends LegacyIntegrationTestCase {
 
 	public function down() {
 		_elgg_services()->session->removeLoggedInUser();
-		$this->user->delete();
+		if ($this->user) {
+			$this->user->delete();
+		}
 		_elgg_services()->hooks->restore();
 	}
